requirements.txt
discord.py>=2.3.0
aot-toolkit
python-dotenv
Pillow>=10.0.0
aiohttp>=3.8.0
